requirements.txt
# PS100 Solar Monitor Dependencies  
# Updated: June 10, 2025
# Optimized for Anker SOLIX PS100 solar panels with TimescaleDB integration

# Core hardware libraries
adafruit-circuitpython-ina228>=1.2.0
adafruit-circuitpython-busdevice>=5.2.0
adafruit-circuitpython-register>=1.9.0
adafruit-blinka>=8.20.0

# Raspberry Pi GPIO (if needed)
RPi.GPIO>=0.7.1
lgpio>=0.2.0

# TimescaleDB and PostgreSQL support
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# SQLite support (fallback)
# sqlite3 is built into Python

# Configuration and data processing
PyYAML>=6.0

# Data processing and statistics
numpy>=1.21.0
pandas>=1.3.0

# Async and web support (future phases)
aiofiles>=23.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
websockets>=11.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.9.0
asyncpg>=0.29.0

# System monitoring
psutil>=5.9.0

# Development tools
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=22.0.0
